fix(dto): corregir DTOs y reorganizar carpetas
