+++ src/modules/users/dto/create-user-verification.dto.ts
@@ -2,6 +2,10 @@ import { IsString, IsOptional, IsEnum, IsNotEmpty } from 'class-validator';
 import { ApiProperty } from '@nestjs/swagger';
 import { VerificationStatus } from '@users/entities/user-verification.entity';
 
+
+
+
+
 export class CreateUserVerificationDto {
   @ApiProperty({
     description: 'URL de la imagen del frente del documento de identidad',
@@ -9,14 +13,14 @@ export class CreateUserVerificationDto {
   @IsString()
   @IsNotEmpty()
   document_front: string;
-
+  
   @ApiProperty({
     description: 'URL de la imagen del reverso del documento de identidad',
   })
   @IsString()
-   @IsNotEmpty()
+  @IsNotEmpty()
   document_back: string;
-
+  
   @ApiProperty({
     description: 'URL de la selfie del usuario para verificación',
   })
@@ -24,7 +28,7 @@ export class CreateUserVerificationDto {
   @IsNotEmpty()
   selfie_image: string;
 
-  @ApiProperty({
+    @ApiProperty({
     description: 'Nota de rechazo opcional',
     required: false,
   })
@@ -33,6 +37,16 @@ export class CreateUserVerificationDto {
   note_rejection?: string;
 }
 
+export class UploadFilesDto {
+  @ApiProperty({ type: 'string', format: 'binary', description: 'Imagen del frente del documento' })
+  document_front: any;
+
+  @ApiProperty({ type: 'string', format: 'binary', description: 'Imagen del reverso del documento' })
+  document_back: any;
+
+  @ApiProperty({ type: 'string', format: 'binary', description: 'Selfie sosteniendo el documento' })
+  selfie_image: any;
+}
 
 export class GetVerificationsQueryDto {
   @IsOptional()
